% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_date_from_period_id}
\alias{add_date_from_period_id}
\title{Add period date column from period ID}
\usage{
add_date_from_period_id(df)
}
\arguments{
\item{df}{dataframe with a period_id column}
}
\value{
dataframe with an extra column called period with the dates
}
\description{
Converts the 2023M02 type and 2023Q2 into 1.2.2023 or 1.4.2023, while
keeping the year alone.
}
