% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{add_rolling_average}
\alias{add_rolling_average}
\title{Add rolling average column to prepared data}
\usage{
add_rolling_average(prep_l, periods = 3, align = "center")
}
\arguments{
\item{prep_l}{prepared list with first element a dataframe with the `value` column}

\item{periods}{number of periods to average}

\item{align}{defaults to "center", "right" and "left" also options}
}
\value{
same list with the df with the `value` column overwritten and the old
values stored in the `raw` column, plus an element with the text for describing
the transformation.
}
\description{
Add a column with a rolling average for a given number of periods. By default
centerred rolling averages are used, but right and left are also possible, see
the documentation for the zoo package for more deets.
}
