% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{add_yoy_change}
\alias{add_yoy_change}
\title{Add year on year change column to prepared data}
\usage{
add_yoy_change(prep_l)
}
\arguments{
\item{prep_l}{prepared list with first element a dataframe with the `value` column}
}
\value{
same list with the df with one the value column overwritten and new element
with  text for describing the transformation and the unit overwritten as well.
}
\description{
Year-on-year change is calculated based on the type of interval and overwrites the
values of the original data. The unit is also changed to "%", adn the
transformation text is added.
}
