% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{all_equal}
\alias{all_equal}
\title{Test that all values in a vector are the same}
\usage{
all_equal(x)
}
\arguments{
\item{x}{vector}
}
\value{
logical
}
\description{
Testing that all values in a vector (or column) are the same.
}
