% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{apply_xlims}
\alias{apply_xlims}
\title{Apply limits to x-axis range of data}
\usage{
apply_xlims(df, xmin = "2011-01-01", xmax = NULL)
}
\arguments{
\item{df}{dataframe with period column in Date format}

\item{xmin}{Date, default "2011-01-01"}

\item{xmax}{Date, default maximum available in data.}
}
\value{
df as input with possibly fewer rows
}
\description{
Takes adataframe with the datapoints and periods and removes the periods before xmin
(default is 1.1.2011) and after xmax (defaults to max in the data).
}
