% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{axis_min_gap}
\alias{axis_min_gap}
\title{Minimum gap between adjacent axis labels, in units of the width of "m"}
\usage{
axis_min_gap()
}
\description{
Minimum gap between adjacent axis labels, in units of the width of "m"
}
\keyword{internal}
