% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calculate_smallest_gap}
\alias{calculate_smallest_gap}
\title{Calculate smallest gap between labels}
\usage{
calculate_smallest_gap(x_positions, x_labels)
}
\arguments{
\item{x_positions}{vector of x positions}

\item{x_labels}{character vector of x labels}
}
\value{
size of smallest gap in "m" width
}
\description{
takes x positions and labels and calculates the smallest gap in the series.
the unit is the width of the letter "m"
}
