% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{center_dates}
\alias{center_dates}
\title{Shift dates to mid-period for proper chart alignment}
\usage{
center_dates(datapoints)
}
\arguments{
\item{datapoints}{list of date+value dataframes}
}
\value{
datapoints with centered dates
}
\description{
Shift dates to mid-period for proper chart alignment
}
\keyword{internal}
