% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{check_and_rename_file}
\alias{check_and_rename_file}
\title{Checks if a file is open/locked and renames it}
\usage{
check_and_rename_file(filename)
}
\arguments{
\item{filename}{file to check}
}
\value{
(updated) filename
}
\description{
If the file you are trying to write to is open, it is also locked for
writing. So in this case we rename it by adding a timestamp before the
extension. Kludgy but works.
}
