% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{check_and_rename_pdf}
\alias{check_and_rename_pdf}
\title{Checks if pdf is open/locked and renames filename}
\usage{
check_and_rename_pdf(filename)
}
\arguments{
\item{filename}{file to check}
}
\value{
(updated) filename
}
\description{
Deprecated, use \link{check_and_rename_file}, which handles any extension.
}
