% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_consistency_or_na}
\alias{check_consistency_or_na}
\title{Check if a column has values and they are all the same}
\usage{
check_consistency_or_na(column)
}
\arguments{
\item{column}{a dataframe column}
}
\value{
any(duplicated(
}
\description{
Column check that returns true if all valid values are the same. This
means all NAs or some NAs are also OK, basically saying the column
has a single unique valid value, which can be used as a parameter downstream
}
\keyword{internal}
