% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{check_plot_inputs}
\alias{check_plot_inputs}
\title{Check input data for multi-line chart}
\usage{
check_plot_inputs(df)
}
\arguments{
\item{df}{input dataframe with at least the following columns: serija, enota}
}
\value{
input df, possibly with updated main titles.
}
\description{
New version for publication ready charts.
Helper function to check that the input data dataframe for multi line charts is
correct. This means that series used in a single chart have to:
- be unique
- have a single unit - possibly two for dual y-axis
- have a maximum of 8 series
}
