% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_uniqueness_or_na}
\alias{check_uniqueness_or_na}
\title{Check if a column has values and they are all the different or NA}
\usage{
check_uniqueness_or_na(column)
}
\arguments{
\item{column}{a dataframe column}
}
\value{
logical
}
\description{
Column check that returns true if all valid values are the different This
means all NAs or some NAs are also OK, no duplicates excep for NAs if they exist
}
\keyword{internal}
