% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{collapse_categories}
\alias{collapse_categories}
\title{Collapse multiple character columns into one}
\usage{
collapse_categories(data)
}
\description{
Collapse multiple character columns into one
}
\keyword{internal}
