% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{colourise}
\alias{colourise}
\title{Colourise text with ANSI true colour}
\usage{
colourise(text, hex)
}
\arguments{
\item{text}{character string to colourise}

\item{hex}{hex colour string}
}
\value{
colourised string or plain text if terminal doesn't support it
}
\description{
Colourise text with ANSI true colour
}
\keyword{internal}
