% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{cut_to_x_range}
\alias{cut_to_x_range}
\title{Cut datapoints to x limits from config}
\usage{
cut_to_x_range(datapoints, config)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}
}
\value{
a list with the xmin and xmax values of the data in date format and the filtered datapoints
}
\description{
takes the maximum range of all the series in the datapoints tables
and then cuts it down by the values of xmin and xmax passed in the
config dictionary. If the intervals don't overlap, uses just the
datapoint range, ignoring the user input.
}
\details{
config$xmax is always the first day of its period (parse_spec_date's
contract), but center_dates() has already shifted plotted dates to
mid-period before this runs. A literal comparison against config$xmax
would therefore always exclude the period xmax names. xmax is rounded
up to the last day of its period before being used as a bound; xmin
is left as-is since mid-period dates are always >= first-of-period.

Returns filtered datapoints.
}
