% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{detect_format}
\alias{detect_format}
\title{Detect whether dataframe is wide or long format}
\usage{
detect_format(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
"wide" or "long"
}
\description{
Detect whether dataframe is wide or long format
}
\keyword{internal}
