% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{determine_interval}
\alias{determine_interval}
\title{Determine interval from timeseries dataframe}
\usage{
determine_interval(df)
}
\arguments{
\item{df}{dataframe with date column}
}
\value{
A, Q or M - or NA
}
\description{
function to return A, Q or M based on differences
in the time periods in the dataframe.
}
