% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{do_transformations}
\alias{do_transformations}
\title{Umbrella fun to do needed transformations on the data}
\usage{
do_transformations(input_data)
}
\arguments{
\item{input_data}{list with dataframe called `data_points`}
}
\value{
list with transformed data points
}
\description{
This funciton takes a list of inputs, which must include the `data_points` and
transformation parameters, and applies the transformations, returning the same list
with transformed data points and the added `transf_txt` element.
}
