% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{draw_areas}
\alias{draw_areas}
\title{Draw area or ribbon shading}
\usage{
draw_areas(datapoints, config, y_axis_label)
}
\arguments{
\item{datapoints}{list of date+value dataframes}

\item{config}{internal config with series info}

\item{y_axis_label}{character y-axis label for baseline detection}
}
\description{
Draws filled polygons for series with type = "area". One area series
shades from baseline (0 by default, 100 for index charts) to the series
values. Two area series shade the region between them.
}
\keyword{internal}
