% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_view_chart.R
\name{draw_axis_break}
\alias{draw_axis_break}
\title{Draw axis break indicator over bottom y-axis label}
\usage{
draw_axis_break(y_lims, bar)
}
\arguments{
\item{y_lims}{numeric(2) y-axis limits}

\item{bar}{logical, is it a bar chart}
}
\description{
Draw axis break indicator over bottom y-axis label
}
\keyword{internal}
