% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_view_chart.R
\name{draw_emphasis}
\alias{draw_emphasis}
\title{Draw emphasis gridlines}
\usage{
draw_emphasis(emphasis, y_axis_label, y_lims)
}
\arguments{
\item{emphasis}{NULL (auto), FALSE (none), or numeric vector}

\item{y_axis_label}{character y-axis label (for auto-detection)}

\item{y_lims}{numeric(2) y-axis limits}
}
\description{
Draw emphasis gridlines
}
\keyword{internal}
