% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{draw_lines}
\alias{draw_lines}
\title{Draw lines onto plot}
\usage{
draw_lines(datapoints, config, x_values = NULL)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}

\item{x_values}{vector of midpoints from \link[UMARvisualisR]{base_barplot}}
}
\value{
nothing, just draws lines
}
\description{
This works for either plotting only lines on a preexisting empty plot
from \link[UMARvisualisR]{empty_plot}, or plotting on top of existing
barplot from \link[UMARvisualisR]{base_barplot} in which case you
have to pass it the bar midpoints as x_values
}
