% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphing.R
\name{dygraph_plotter}
\alias{dygraph_plotter}
\title{Dygraph line charts}
\usage{
dygraph_plotter(prep_l)
}
\arguments{
\item{prep_l}{list of length 8+ with data.frame with data, the unit used and the
main and sub titles etc.. see \link[UMARvisualisR]{prep_multi_line}.}
}
\value{
nothing, plots to open device
}
\description{
Plots univariate and multivariate line chart to dygraph interactive charts,
approx in line with our corporate identity.
}
\details{
Data input must be prepared with the \link[UMARvisualisR]{prep_multi_line} function.
}
