% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphing.R
\name{dygraph_plotter_mixed}
\alias{dygraph_plotter_mixed}
\title{Dygraph bar and combo charts}
\usage{
dygraph_plotter_mixed(prep_l, bar = TRUE)
}
\arguments{
\item{prep_l}{list of length 8+ with data.frame with data, the unit used and the
main and sub titles etc.. see \link[UMARvisualisR]{prep_multi_line}.}

\item{bar}{logical if first series is plotted as a bar and the rest line or vice versa.}
}
\description{
Plots univariate barchart and multivariate combo chart to dygraph interactive charts,
approx in line with our corporate identity.
}
\details{
If only a single series is passed to the funciton it will automaticlaly print
a barchart

if more are passed the default is that the first series becomes a barchart and the
rest are line (bar = TRUE), alternatively if bar == FALSE that means the first one
is a line chart and the rest are a stacked barchart. In both cases a single y-axis
is used.

Data input must be prepared with the \link[UMARvisualisR]{prep_multi_line} function.
}
