% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{emf_output}
\alias{emf_output}
\title{Output to emf}
\usage{
emf_output(filename, size = "normal")
}
\arguments{
\item{filename}{name of file with path}

\item{size}{small, normal or large, defaults to normal}
}
\value{
nothing, produces emf file
}
\description{
Thin wrapper on \link{chart_device}. Needs the \pkg{devEMF} package.
EMF is Word's native vector format on Windows.
}
