% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{empty_plot}
\alias{empty_plot}
\title{Plots empty plot with gridlines}
\usage{
empty_plot(x_lims, y_axis, y_axis_label, pad = c(0, 0.01))
}
\arguments{
\item{x_lims}{output of \link[UMARvisualisR]{get_x_lims}}

\item{y_axis}{output of \link[UMARvisualisR]{find_pretty_ylim}}

\item{y_axis_label}{whatever is in config$y_axis_label - required to
check if the unit is an index, so the 100 gridline can be emphasised.}

\item{pad}{numeric or vector of length 2 padding x axis ednes.}
}
\value{
nothing, plots to open device
}
\description{
Plots an empty plot with the gridlines, emphasised on 100 for indices a
and zero if in range.
}
