% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{end_controlled_plot}
\alias{end_controlled_plot}
\title{Close plot for testing}
\usage{
end_controlled_plot(temp_file)
}
\value{
nothing, closes the png device
}
\description{
Use temp_file <- start_controlled_plot() and then close and unlink with this
}
\keyword{internal}
