% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_na_labels}
\alias{filter_na_labels}
\title{Filter NA labels and positions}
\usage{
filter_na_labels(x_positions, x_labels)
}
\arguments{
\item{x_positions}{vector of label positions}

\item{x_labels}{vector of labels}
}
\value{
list with both filtered vectors
}
\description{
Taking the equal lenght label and position vector, removes the indices in
both vectors  where the labels are NA.
}
