% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{find_date_column}
\alias{find_date_column}
\title{Find the first Date-class column}
\usage{
find_date_column(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
column name or NULL
}
\description{
Find the first Date-class column
}
\keyword{internal}
