% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{find_pretty_ylim}
\alias{find_pretty_ylim}
\title{Find pretty y-axis limits}
\usage{
find_pretty_ylim(values)
}
\arguments{
\item{values}{vector of numeric values}
}
\value{
numeric vector of length 2 to be used as y-axis limit
}
\description{
Finds y/limits that lead to a pretty y-axis. If the range of the values is
on the line, it artificially extends it by one more break.
}
