% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{first_up}
\alias{first_up}
\title{Helper function to capitalise firs letter of string}
\usage{
first_up(x)
}
\arguments{
\item{x}{character string}
}
\value{
a character string with first letter capitalized.
}
\description{
Capitalise the first letter of a string, but keep capitalization as is
on other letters.
}
