% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{format_number}
\alias{format_number}
\title{Format numbers for axis labels}
\usage{
format_number(x, language = "si")
}
\arguments{
\item{x}{numeric vector}

\item{language}{"si" or "en"}
}
\description{
Format numbers for axis labels
}
\keyword{internal}
