% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_output_chart.R
\name{get_code}
\alias{get_code}
\title{Generate reproducible code for a chart}
\usage{
get_code(chart)
}
\arguments{
\item{chart}{an object of class "umar_chart" from \link[UMARvisualisR]{prep_chart}}
}
\value{
character string with the code (also printed to console)
}
\description{
Returns the full \code{prep_chart()} call with all parameters
explicitly specified, so the chart can be reproduced or modified.
}
