% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{get_data}
\alias{get_data}
\title{Get the datapoints from the database}
\usage{
get_data(config, con)
}
\arguments{
\item{config}{configuration list}

\item{con}{connection to database}
}
\value{
list with a df for each series
}
\description{
taking the configuration list output from \link[UMARvisualisR]{prep_config} returns
a list with a dataframe for each of the series of the correct vintage
}
