% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_data_values}
\alias{get_data_values}
\title{Get data values as a vector}
\usage{
get_data_values(datapoints, config)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}
}
\value{
vector of numeric values
}
\description{
Returns the data values as a vector to be used to get the ylims, but
also adds a 0 to the range if the chart is a barchart and also makes
sure the ylims are correct if it's a stacked barchart. Christ, i am
really overdoing this shit.
}
