% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_date_from_period}
\alias{get_date_from_period}
\title{New fun to get date from period with three types of alignment}
\usage{
get_date_from_period(period, position = c("c", "l", "r"))
}
\arguments{
\item{period}{one of three types}

\item{position}{l, c or r, defaults to middle}
}
\value{
date
}
\description{
takes year, quarter or month format and returns first ("l"), middle ("c")
or last ("r") date, the default is "c"
}
