% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_interval_in_months}
\alias{get_interval_in_months}
\title{Helper to get the time interval from a date column}
\usage{
get_interval_in_months(date_column)
}
\arguments{
\item{date_column}{what it says on the box}
}
\value{
integer value in months
}
\description{
Helper to get the time interval from a date column
}
