% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{get_legend_labels_from_df}
\alias{get_legend_labels_from_df}
\title{Get legend labels from input dataframe}
\usage{
get_legend_labels_from_df(df, original_table_names = NULL)
}
\arguments{
\item{df}{input dataframe with at least the following columns: `series_name`, `chart_no`}

\item{original_table_names}{character vector with original table names, which is only relevant
if series names are identical.}
}
\value{
character vector of length nrow(df) with legend labels and optionally a second one
with the new chart name.
}
\description{
One of the columns is the `series_name` one, which is from the `series` table and contains
the dimension values separated by `--`. If these names have the same number of dimension
values and differ by just one of them, then that one is returned as the legend labels.
Warnings are issued for different numbers of dimensions or differences in more than
one dim, or if the series names are the same and no table names are provided,
in which three cases NAs are returned.
}
\details{
If the series are different by one dimension, that is returned, and if the series
are the same and the different table names are provided, then those are returned.
}
