% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_legend_lines}
\alias{get_legend_lines}
\title{helper function to get number of lines required for the legend}
\usage{
get_legend_lines(elements, columns)
}
\arguments{
\item{elements}{string of legend entries}

\item{columns}{numeric value of number of columns}
}
\value{
integer value
}
\description{
helper function to get number of lines required for the legend
}
