% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_max_period}
\alias{get_max_period}
\title{Get max period id from list of dataframes with with periods and their ids}
\usage{
get_max_period(data_points)
}
\arguments{
\item{data_points}{list of dataframes with period and period_id columns}
}
\value{
value of the most recent period_id in list
}
\description{
Helper function to get the period id for the most recent data point in a
list of dataframes which have period (date) and period_id columns (i.e. 2023M03)
}
\keyword{internal}
