% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_most_recent_dataframe}
\alias{get_most_recent_dataframe}
\title{Get dataframe with most recent datapoint}
\usage{
get_most_recent_dataframe(datapoints)
}
\arguments{
\item{datapoints}{list of dataframes with date column}
}
\value{
single dataframe with date column
}
\description{
Takes list of dataframes and returns the one with the most recent date value
}
