% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_most_recent_interval}
\alias{get_most_recent_interval}
\title{Get the interval of the most recent dataframe datapoint}
\usage{
get_most_recent_interval(datapoints)
}
\arguments{
\item{datapoints}{list of dataframes with date column}
}
\value{
"A", "Q" or "M"
}
\description{
Takes list of dataframes with date column and returns A, Q or M
}
