% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{interpolate_x}
\alias{interpolate_x}
\title{Linear interpolation for x axis transformation}
\usage{
interpolate_x(original_dates, x_values, new_dates)
}
\arguments{
\item{original_dates}{date vector same length as x_values}

\item{x_values}{numeric vector, same length as orinal dates}

\item{new_dates}{date vector}
}
\value{
interpolated x values of new dates
}
\description{
used to transform a numeric axis to a date axis and scale other
date vectors to the axis coordinates. assumes the input dates and
values are on a line i.e. only takes the first and last values into
account
}
