% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last_year_complete}
\alias{last_year_complete}
\title{Check if at least one timeseries has a complete last year}
\usage{
last_year_complete(datapoints)
}
\arguments{
\item{datapoints}{list od datapoints dataframes}
}
\value{
logical
}
\description{
Takes a list of dataframes with a date column and checks
if at least one of the series has a complete final year.
}
