% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last_year_complete_series}
\alias{last_year_complete_series}
\title{Check if timeseries has a complete last year}
\usage{
last_year_complete_series(datapoints_df)
}
\arguments{
\item{datapoints_df}{dataframe with date column}
}
\value{
logical
}
\description{
Check if timeseries has a complete last year
}
