% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{legend_mz2}
\alias{legend_mz2}
\title{My legend function}
\usage{
legend_mz2(
  x = par("usr")[[1]],
  y = par("usr")[[4]],
  legend,
  fill = NULL,
  col = NULL,
  lty = NULL,
  lwd = NULL,
  xjust = 0,
  yjust = 1,
  x.intersp = 0.2,
  y.intersp = 0.7,
  text.col = "black",
  text.font = NULL,
  ncol = 1,
  family = "Myriad Pro"
)
}
\arguments{
\item{x}{the x and y co-ordinates to be used to position the legend. They can be specified by keyword or in any way which is accepted by xy.coords: See ‘Details’.}

\item{y}{the x and y co-ordinates to be used to position the legend. They can be specified by keyword or in any way which is accepted by xy.coords: See ‘Details’.}

\item{legend}{a character or expression vector of length ≥ 1 to appear in the legend. Other objects will be coerced by as.graphicsAnnot.}

\item{fill}{if specified, this argument will cause boxes filled with the specified colors (or shaded in the specified colors) to appear beside the legend text.}

\item{col}{color of points or lines appearing in the legend.}

\item{lty}{the line types and widths for lines appearing in the legend. One of these two must be specified for line drawing.}

\item{lwd}{the line types and widths for lines appearing in the legend. One of these two must be specified for line drawing.}

\item{xjust}{how the legend is to be justified relative to the legend x location. A value of 0 means left justified, 0.5 means centered and 1 means right justified.}

\item{yjust}{the same as xjust for the legend y location.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing.}

\item{y.intersp}{the same for vertical (y) line distances.}

\item{text.col}{the color used for the legend text.}

\item{text.font}{the font used for the legend text, see text.}

\item{ncol}{the number of columns in which to set the legend items (default is 1, a vertical legend).}

\item{family}{font family}
}
\value{
nothing, draws legend
}
\description{
My legend function
}
