% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{linestyle_to_lty}
\alias{linestyle_to_lty}
\title{Convert linestyle to lty}
\usage{
linestyle_to_lty(style)
}
\arguments{
\item{style}{auto passed}
}
\value{
nothing, closes the png device
}
\description{
Internal function
}
\keyword{internal}
