% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{multi_checks}
\alias{multi_checks}
\title{Check input data for multi-line chart}
\usage{
multi_checks(df)
}
\arguments{
\item{df}{input dataframe with at least the following columns: code, unit_name
interval_id, rolling_average_alignment, rolling_average_periods, year_on_year}
}
\value{
input df, possibly with updated main titles.
}
\description{
Helper function to check that the input data dataframe for multi line charts is
correct. This means that series used in a single chart have to be:
- multiple
- have the same unit
- have the same interval (this may be changed in the future)
- are either all rolling averages or not and if they are, they have the same
alignment and number of periods (might also change in the future)
- are either all y-o-y changes or not (might also change in the future).
If any of these is true, the function stops
}
\details{
Also checks if
- main_titles are the same, otherwise it deletes them, but only issues a warning.
}
