% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{multi_titles}
\alias{multi_titles}
\title{Check titles in input for multi-line chart}
\usage{
multi_titles(df)
}
\arguments{
\item{df}{}
}
\description{
Check if the main_title is the same for all series, and if not issue a warning and
remove all the titles
}
