% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrellas.R
\name{multiline_pipeline}
\alias{multiline_pipeline}
\title{Multiline pipeline from input dataframe}
\usage{
multiline_pipeline(df, con, xmin = "2011-01-01", xmax = NULL)
}
\arguments{
\item{df}{dataframe with the following fields: `tabel_name`, `series_name`, `series_code`,
`unit_name`,    `interval_id`,    `chart_no`,    `rolling_average_periods`,
`rolling_average_alignment`,    `year_on_year`.}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
plots chart
}
\description{
Pipeline that takes an appropriate dataframe with instrucitons for the charts and
prepares the data and outputs the the charts. SHould work on single and multiline
and on transformed data as well as long as all the series in a multiline have the
same transformation.
}
