% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{multivariate_line_chart}
\alias{multivariate_line_chart}
\title{Framework for plotting a multivariate line chart}
\usage{
multivariate_line_chart(prep_l, xmin = "2011-01-01", xmax = NULL)
}
\arguments{
\item{prep_l}{list of length 8+ with data.frame with data, the unit used and the
main and sub titles etc.. see \link[UMARvisualisR]{prep_multi_line}.}

\item{...}{xmin or xmax to be passed to \link[UMARvisualisR]{apply_xlims}}
}
\value{
nothing, plots to open device
}
\description{
Plots univariate and multivariate line chart in line with the UMAR corporate
graphical identity.
Data input must be prepared with the \link[UMARvisualisR]{prep_multi_line} function.
At the moment only plots years as labels.
}
\details{
Todo: add arguments such as x-value range, add flexibility of axis labels for different
time ranges..
}
