% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{na_chart}
\alias{na_chart}
\title{Plotting function for chart with no data values}
\usage{
na_chart(prep_l)
}
\arguments{
\item{prep_l}{list of length four with data.frame with data, the unit used and the
main and sub titles. see \link[UMARaccessR]{prep_single_line}.}
}
\value{
nothing, plots to open device
}
\description{
Plots error chart with message when all the data values in the chart are
missing. To be used as a temporary placeholder in reports where the selection
accidentally doens't include any data. Input data must be prepared by
\link[UMARaccessR]{prep_single_line}. Called from  \link[UMARvisualisR]{univariate_line_chart}
if all values are NA.
}
