% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{only_annual_intervals}
\alias{only_annual_intervals}
\title{Check if all datapoints have annual data only}
\usage{
only_annual_intervals(datapoints)
}
\arguments{
\item{datapoints}{list od datapoints dataframes}
}
\value{
logical value indicating truth
}
\description{
Takes list of dataframes with datapoints and checks that all of them
have annual data only - regardless of the horizontal alignment i.e.
start, middle or end of they year.
}
