% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{parse_long}
\alias{parse_long}
\title{Parse long format into standard internal structure}
\usage{
parse_long(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
list with datapoints (list of dfs) and series_names
}
\description{
Parse long format into standard internal structure
}
\keyword{internal}
