% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{parse_wide}
\alias{parse_wide}
\title{Parse wide format into standard internal structure}
\usage{
parse_wide(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
list with datapoints (list of dfs) and series_names
}
\description{
Parse wide format into standard internal structure
}
\keyword{internal}
