% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{prep_config}
\alias{prep_config}
\title{Prepare plot configuration from dataframe}
\usage{
prep_config(df, con)
}
\arguments{
\item{df}{with at least the columns series_code and unit}

\item{con}{database connection}
}
\value{
a config list for plotting
}
\description{
Takes the input dataframe which may or may not have config data for the plot
and the series and should have gone through the \link[UMARvisualisR]{check_plot_inputs}
check first!
Updates the default config with whatever is in the dataframe. Defaults are
hardcoded here and the prep will work even if only the series codes are
given in the dataframe, which is the bare minimum and will produce a basic
line plot configuration. The output config is the input to the plotting function.
}
