% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{prep_config_en}
\alias{prep_config_en}
\title{Change axis label for english version}
\usage{
prep_config_en(config)
}
\arguments{
\item{config}{configuration list}
}
\value{
updated config list
}
\description{
used inside \link[UMARvisualisR]{publication_ready_plot} to change
y axis label to english version.
}
