% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{prep_data}
\alias{prep_data}
\title{Whole data prep pipeline for plotting publicaiton ready charts}
\usage{
prep_data(df, con, schema = "platform")
}
\arguments{
\item{df}{input dataframe with at least the following columns: serija, enota}

\item{con}{database connection}

\item{schema}{character schema name, default is "platform"}
}
\value{
list of dataframes and config list
}
\description{
takes as its input a dataframe with one series per row and all the relevant info.
(see vignette that hasn't been written yet). Checks the data input is OK,
prepares the config, gets the data, transforms the data if necessary and
returns a list od dataframes with all the datapoints and the config list.
}
