% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{prep_multi_line}
\alias{prep_multi_line}
\title{Prepare data needed for multi (or single) line chart}
\usage{
prep_multi_line(df, con, date_valid = NULL)
}
\arguments{
\item{df}{dataframe described above}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{date_valid}{validity of vintages, NULL gives most recent.}
}
\value{
A list of lentgh seven described above.
}
\description{
Uses an input table which must have the following columns: `itnerval_id`,
`unit_name`, `main_title`, `name_long` for the subtitle, `id` for the series id
(same as gets input into \link[UMARvisualisR]{multi_checks}, which is
run on the dataframe as the first step in this funciton, to check everything is cool with
the inputs.
}
\details{
Prepares the list of data, structured to get plotted properly. Currently seven
elements are prepared: `data_points` a list of dataframes with the actual data,
`unit`, `main_title` , `sub_title`, `updated`, `last_period`, `interval`.

Needs a valid connection to get the data as well as the last published date.
}
