% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prep_single_line}
\alias{prep_single_line}
\title{Prepare data needed for a univariate line chart}
\usage{
prep_single_line(vintage, con)
}
\arguments{
\item{vintage}{numeric id of vintage}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
a list with the data frame with values, period_ids and periods as the
first element, a character unit name as second, and the row wrapped main and
subtitles (default 100 chars max 3 lines), the date and time of the last update,
the last period and the interval.
}
\description{
Given the vintage id and a connection to the database, this function
gets the datapoints, and the unit, prepares the titles. The returned list
is the input for the plotting function \link[UMARvisualisR]{univariate_line_chart}.
}
