% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_prep_chart.R
\name{print.umar_chart}
\alias{print.umar_chart}
\title{Print method for umar_chart objects}
\usage{
\method{print}{umar_chart}(x, ...)
}
\arguments{
\item{x}{umar_chart object}

\item{...}{ignored}
}
\description{
Print method for umar_chart objects
}
