% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{publication_ready_plot}
\alias{publication_ready_plot}
\title{Umbrella function for publicaiton ready chart}
\usage{
publication_ready_plot(datapoints, config, language = "si")
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}

\item{language}{"si" or "en"}
}
\description{
Umbrella function for publicaiton ready chart
}
