% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrellas.R
\name{publication_ready_plot_pipeline}
\alias{publication_ready_plot_pipeline}
\title{Full pipeline for publicaiton ready charts}
\usage{
publication_ready_plot_pipeline(input_file)
}
\arguments{
\item{input_file}{excel with single sheet and all relevant info for config.
see vingette (non existent as of now) for full details.}
}
\value{
nothing, produces pdf charts
}
\description{
Input appropriate excel file with config info and produces an english and
slovenian version of the chart in pdf format, currently to the hardcoded
folder "O:\\Avtomatizacija\\umar-publication-ready-grafi\\grafi\\",
but this can of course be changed if needed.
}
