% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{quarterly_label}
\alias{quarterly_label}
\title{Format date into quarterly label}
\usage{
quarterly_label(date)
}
\arguments{
\item{date}{date}
}
\value{
character vector of quarterly labels
}
\description{
Q1-23 format
}
