% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{in_range}
\alias{in_range}
\alias{in_range_strict}
\title{Helper functions for testing if a value is in a range}
\usage{
in_range(x, r)

in_range_strict(x, r)
}
\arguments{
\item{x}{numeric vector of length one}

\item{r}{a range - numeric vector of lenght two, ascending}
}
\value{
logical
}
\description{
in_range checks if the value is within the range, including the limits,
in_range_strict excludes the limits.
}
