% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_csv_guess_encoding}
\alias{read_csv_guess_encoding}
\title{Helper function to guess encoding of a csv file and read slovenian csv}
\usage{
read_csv_guess_encoding(file)
}
\arguments{
\item{file}{path to csv file}
}
\value{
content of the csv file
}
\description{
guesses encoding and reads csv with slovenian delimiters
}
