% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{remove_head_tail_NAs}
\alias{remove_head_tail_NAs}
\title{Remove NA rows at start and end}
\usage{
remove_head_tail_NAs(df)
}
\arguments{
\item{df}{data frame with at least period and value columns}
}
\value{
dataframe with possibly fewer rows than going in.
}
\description{
Removes rows containing NAs in the column `value`), but
only at the beginning and end of the table.
}
