% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{replace_period_id_column}
\alias{replace_period_id_column}
\title{Replace the period_id column with the correctly alligned date}
\usage{
replace_period_id_column(df, position = "c")
}
\arguments{
\item{df}{dataframe with period_id column}

\item{position}{defaults to "c", but "l" and "r" are also valid}
}
\value{
df with same number of cols
}
\description{
Replace the period_id column with the correctly alligned date
}
