% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rolling_average}
\alias{rolling_average}
\title{Add rolling average column}
\usage{
rolling_average(df, periods = 3, align = "center")
}
\arguments{
\item{df}{dataframe with the `value` column}

\item{periods}{number of periods to average}

\item{align}{defaults to "center", "right" and "left" also options}
}
\value{
a df with one more column named `rolling`
}
\description{
Add a column with a roling average for a given number of periods. By default
centered rolling averages are used, but right and left are also possible, see
the documentation for the zoo package for more deets.
}
