% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rounding}
\alias{rounding}
\title{Round like Excel}
\usage{
rounding(x, digits = 0)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits}
}
\value{
vector of same lenght as x
}
\description{
Round so that .5 is always rounded up, unlike `base::round`
}
