% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{shift_dates_by_six_months}
\alias{shift_dates_by_six_months}
\title{Add 6 months to date vector}
\usage{
shift_dates_by_six_months(dates)
}
\arguments{
\item{dates}{vector of dates}
}
\value{
vector of same length shifted to the right by 6 months
}
\description{
Add 6 months to date vector
}
