% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{split_by_unit}
\alias{split_by_unit}
\title{Function to split the datapoints into left and right axis.}
\usage{
split_by_unit(datapoints, config)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}
}
\value{
list of left and right datapoints and configs (if right exists)
}
\description{
Function to split the datapoints into left and right axis.
}
