% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{start_controlled_plot}
\alias{start_controlled_plot}
\title{Open plot for testing}
\usage{
start_controlled_plot(width = 800, height = 600)
}
\arguments{
\item{width}{width in px}

\item{height}{height in px}
}
\description{
Open plot for testing
}
\keyword{internal}
