% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{strip_na_rows}
\alias{strip_na_rows}
\title{Strip rows with NA values}
\usage{
strip_na_rows(df)
}
\arguments{
\item{df}{dataframe with date and value column}
}
\value{
df with same number of cols
}
\description{
Strip rows with NA values
}
