% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{transform_data}
\alias{transform_data}
\title{Transform a series dataframe with the series confing}
\usage{
transform_data(df, series_config)
}
\arguments{
\item{df}{dataframe with date and value columns}

\item{series_config}{list of series' config paramters from
\link[UMARvisualisR]{prep_config}.}
}
\value{
df and series' config
}
\description{
Takes the dataframe with date and value columns and the series'
config list (not the chart config list, just the series') and
does the transformations. Currently the default is that rolling
averages are performed first, but i can change that later and
add the other way round. The \link[UMARvisualisR]{check_plot_inputs}
function has already made sure the combinations of transformations
are all legal.
}
