% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_growth}
\alias{transform_growth}
\title{New function for growth transformations}
\usage{
transform_growth(df, type = "YOY")
}
\arguments{
\item{df}{dataframe with date and value columns}

\item{type}{valid are YOY, QOQ and MOM}
}
\value{
df with updated valeus
}
\description{
New function for growth transformations
}
