% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_index}
\alias{transform_index}
\title{FUnction for transformation to an index with a specific base period}
\usage{
transform_index(df, base_period)
}
\arguments{
\item{df}{data frame with date and value columns}

\item{base_period}{either a year in YYYY format or quarter in 2023Q2 format
or month in 2023M03 format}
}
\value{
dataframe with trnsformed values
}
\description{
Calculates the intex based on the period. The following options are allowed:
}
\details{
+ the base period is a year: the average will be taken of the months or quarters,
of course this works fine for annual data as well.
+ if the base period is a quarter, you can only do the transformation on
quarterly data - but this is checked in the check_plot_inputs function
+ if the base period is a month, you can also only use it on monthly data.
}
