% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_rolling}
\alias{transform_rolling}
\title{New function for rolling average transformations}
\usage{
transform_rolling(df, periods = 3, align = "c")
}
\arguments{
\item{df}{dataframe with date and value columns}

\item{periods}{number of periods to be averaged over}

\item{align}{alignment method - center is the default, right is correct and
left i don't know why it even exists}
}
\value{
dataframe with transformed values.
}
\description{
New function for rolling average transformations
}
