% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{umar_cols}
\alias{umar_cols}
\title{UMAR colour palette function}
\usage{
umar_cols(...)
}
\arguments{
\item{...}{optional colour names.}
}
\value{
hex value of relevant colour(s)
}
\description{
Returns colour(s) from the UMAR corporate graphical identity palette.
Colour names are rdeca, roza, siva, temno_modra, turkizna, zelena, vijolicna,
sinja, gridlines and emph. The last two are grays for gridlines, axes and the like
and for emphasising the 100/0 gridline.
}
