% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{umar_font}
\alias{umar_font}
\title{Get or set the font family used in UMAR charts}
\usage{
umar_font(family = NULL)
}
\arguments{
\item{family}{character font family name. If NULL, returns current value.}
}
\value{
current font family (invisibly when setting)
}
\description{
Get or set the font family used in UMAR charts
}
