% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{umar_palette}
\alias{umar_palette}
\title{Get or set the active UMAR colour palette}
\usage{
umar_palette(palette)
}
\arguments{
\item{palette}{"default" or "Janez" (or "janez"). Omit to get current value.}
}
\value{
active palette name (invisibly when setting)
}
\description{
The default is the UMAR-style palette used in all official publications.
The alternative is the Janez-style colour blindness palette. Set it once per
session!
}
