% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unique_without_na}
\alias{unique_without_na}
\title{Get unique values without NA}
\usage{
unique_without_na(vec)
}
\arguments{
\item{vec}{vector}
}
\value{
vector of unique without NAs
}
\description{
Get unique values without NA
}
\keyword{internal}
