% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{univariate_full}
\alias{univariate_full}
\title{Full pipeline for univariate chart}
\usage{
univariate_full(series, con, date_valid = NULL)
}
\arguments{
\item{series}{numeric id of series}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{date_valid}{date when the vintage was valid if none is given most recent
i.e. currently valid vintage is returned.}
}
\description{
From the series id and the connection, gets the appropriate vintage id,
then the data prepared for the chart and finally plots the chart.
}
