% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{univariate_line_chart}
\alias{univariate_line_chart}
\title{Plotting function for univariate line chart}
\usage{
univariate_line_chart(prep_l, ...)
}
\arguments{
\item{prep_l}{list of length four with data.frame with data, the unit used and the
main and sub titles. see \link[UMARaccessR]{prep_single_line}.}

\item{xmin}{minimum of x-axis date range}
}
\value{
nothing, plots to open device
}
\description{
Plots univariate line chart in line with the UMAR corporate graphical identity.
Data input must be prepared with the \link[UMARaccessR]{prep_single_line}. At the
moment only plots years as labels.
}
\details{
Todo: add arguments such as x-value range, add flexibility of labels for different
time ranges..
}
