% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umbrellas.R
\name{univariate_line_pipeline}
\alias{univariate_line_pipeline}
\title{FUll pipeline function for plotting a univariate line chart}
\usage{
univariate_line_pipeline(
  series,
  date_valid = NULL,
  xmin = "2011-01-01",
  xmax = NULL,
  rolling = FALSE,
  roll_periods = 3,
  roll_align = "center",
  yoy = FALSE,
  interval = NULL,
  unit = NULL,
  main_title = NULL,
  sub_title = NULL,
  con
)
}
\arguments{
\item{series}{numeric id of series}

\item{date_valid}{date when the vintage was valid if none is given most recent
i.e. currently valid vintage is returned.}

\item{xmin}{character date, default "2011-01-01"}

\item{xmax}{character date, default today}

\item{rolling}{logical / should a rolling average be calculated and displayed}

\item{roll_periods}{number of periods to roll over (defualt = 3)}

\item{roll_align}{alignment of rolling average (default = "center")}

\item{yoy}{logical - should the year on year change be calculated and displayed}

\item{interval}{character describing interval type}

\item{unit}{character describing unit to be written as y label}

\item{main_title}{character..}

\item{sub_title}{character...}

\item{con}{PostgreSQL connection object created by the RPostgres package.}
}
\value{
plots a univariate line chart
}
\description{
From the series id and the connection, gets the appropriate vintage id,
then the data prepared for the chart and finally plots the chart.
}
