% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{update_units}
\alias{update_units}
\title{Updates units in publication chart input table}
\usage{
update_units(df, con)
}
\arguments{
\item{df}{with at a minimum the columns serija, enota, indeks_let and rast}

\item{con}{database connection}
}
\value{
df with updated
}
\description{
Updates the units based on the transformations - to index or% if necessary -
and then queries the database for the remaining units if there are still
any missing. And fixes the capitalisation ready for the y-label
}
