% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_view_chart.R
\name{view_chart}
\alias{view_chart}
\title{Plot a prepared chart to the current device}
\usage{
view_chart(chart)
}
\arguments{
\item{chart}{an object of class "umar_chart" from \link[UMARvisualisR]{prep_chart}}
}
\value{
invisible chart object (for piping)
}
\description{
Takes a chart object from \link[UMARvisualisR]{prep_chart} and renders it
to the current graphics device.
}
