% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{wrap_string}
\alias{wrap_string}
\title{Wrap a character string to a specified character width}
\usage{
wrap_string(string, width)
}
\arguments{
\item{string}{a character string to be wrapped- vector of length one}

\item{width}{numeric - desired width}
}
\value{
a character sting with new lines if required.
}
\description{
Uses strwrap to split a string into subsets that are shorter than the
specified width and inserts newlines in between them. Inspired by this
[SO answer](https://stackoverflow.com/questions/7367138/text-wrap-for-plot-titles/7367534#7367534).
}
