% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wrap_title}
\alias{wrap_title}
\title{Helper funciton to wrap title to appropriate width}
\usage{
wrap_title(title, max_width = NULL, cex = 1, family = "Myriad Pro", font = 1)
}
\arguments{
\item{title}{character}

\item{max_width}{in user units - default is plot width from par("usr")}

\item{cex}{numeric}

\item{family}{of fonts.}

\item{font}{value for font style}
}
\value{
title but wiht line breaks where needed
}
\description{
Had chatgpt write this for me pretty much. does what it says on the box.
}
