% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{write_metadata}
\alias{write_metadata}
\title{Write metadata to first sheet of workbook}
\usage{
write_metadata(prep_l, sheet_name, df, wb, odd)
}
\arguments{
\item{prep_l}{prepared data}

\item{sheet_name}{character string with chart number, also used as sheet name}

\item{wb}{workbook object}
}
\value{
nothing, side effect is writing to excel
}
\description{
Prepares the metadata to be written to the first sheet in the workbook
and appended each time to the stuff already there.
}
\keyword{internal}
