% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{write_to_sheet}
\alias{write_to_sheet}
\title{Write prepared data to an excel spreadsheet}
\usage{
write_to_sheet(prep_l, df, wb, odd)
}
\arguments{
\item{prep_l}{list output from \link[UMARvisualisR]{rename_columns}}

\item{df}{dataframe input of series for each chart - needed to get the chart number}

\item{wb}{an active workbook binding to write to.}

\item{odd}{logical indicating alternating charts, used in the TOC of the Excel
spreadsheet for the backgrounds for greater readability.}
}
\value{
side effect is writing to the excel file.
}
\description{
Taking the prepared data with the columns renamed by \link[UMARvisualisR]{rename_columns}
and joining into a single dataframe, writing it to a sheet and formatting a bit.
}
