% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{x_axis_label_params}
\alias{x_axis_label_params}
\title{Get x axis parameters}
\usage{
x_axis_label_params(datapoints, config, tickmarks, x_lims)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{dictionary list from \link[UMARvisualisR]{prep_config}}

\item{tickmarks}{vector from from \link[UMARvisualisR]{x_axis_lims_tickmarks}}

\item{x_lims}{vector from from \link[UMARvisualisR]{x_axis_lims_tickmarks}}
}
\value{
list of named x_positions, x_labels, tickmars and x_lims
}
\description{
Takes the datapoints and config (for user xmin and xmax mainly) and
calculates the ticmkark positions, the label positions, the default
labels (not handling overlapping just yet) and the xlims for the x axis.
}
\details{
Taking into account if the data is only annual - in which case the tickmarks
and lables align - or monthly/quarterly in which case the labels are between.
Taking into account if the last year is complete or not, which determines
the last tickmark and label etc.
}
