% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{x_axis_lims_tickmarks}
\alias{x_axis_lims_tickmarks}
\title{Get x_axis lims and tickmarks}
\usage{
x_axis_lims_tickmarks(datapoints, config)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{dictionary list from \link[UMARvisualisR]{prep_config}}
}
\value{
named list of tickmarks and x_lims
}
\description{
Get x_axis lims and tickmarks
}
