% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{year_squisher}
\alias{year_squisher}
\title{Year label squishing function}
\usage{
year_squisher(sequence, extra = FALSE)
}
\arguments{
\item{sequence}{numeric sequence of YYYY values of lenght >2}

\item{extra}{logical indicator of extra squishing}
}
\value{
same length sequence of character labels
}
\description{
Squishes year labels using medium squishing - all but the first label
go to YY - or extra squishing: keeping only every 5th one, plus the first
and last value.
}
