% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{year_squisher_extra}
\alias{year_squisher_extra}
\title{Extra  year label squishing}
\usage{
year_squisher_extra(sequence)
}
\arguments{
\item{sequence}{numeric sequence of YYYY values of lenght >2}
}
\value{
same length sequence of character labels
}
\description{
Squishes all but the first year to YY instead of YYYY, but
only keeps every 5th year plust the first and last value.
}
