% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{year_squisher_medium}
\alias{year_squisher_medium}
\title{Medium year label squishing}
\usage{
year_squisher_medium(sequence)
}
\arguments{
\item{sequence}{numeric sequence of YYYY values of lenght >2}
}
\value{
same length sequence of character labels
}
\description{
Squishes all but the first year to YY instead of YYYY
}
