% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{add_yoy_of_rolling}
\alias{add_yoy_of_rolling}
\title{Add Year on year change to a rolling average}
\usage{
add_yoy_of_rolling(data_points, periods = 3, align = "center", interval)
}
\arguments{
\item{data_points}{prepared list of dataframes with the `value` column}

\item{periods}{number of periods to average}

\item{align}{defaults to "center", "right" and "left" also options}

\item{interval}{char id of interval e.g. "M"}
}
\value{
same list with the df with one the value column overwritten and new element
with  text for describing the transformation and the unit overwritten as well.
}
\description{
Uses both transformations: first a rolling average over a set number of periods
(defaults to 3) and then calculates the year on year  change of that average.
Also adds the transformation text to be used in the plotting.
}
