% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{base_barplot}
\alias{base_barplot}
\title{Base barplot}
\usage{
base_barplot(datapoints, config, y_axis)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}

\item{y_axis}{output of \link[UMARvisualisR]{find_pretty_ylim}}
}
\value{
midpoints, of course also plots a barplot to open device
}
\description{
Can be used as base for combo of bar and line plot as well.
All the info is in the datapoints and config.
}
