% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{check_and_rename_pdf}
\alias{check_and_rename_pdf}
\title{Checks if pdf is open/locked and renames filename}
\usage{
check_and_rename_pdf(filename)
}
\arguments{
\item{filename}{file to check}
}
\value{
(updated) filename
}
\description{
if the pdf you are trying to write to is open, it is also locked for
writing. so in this case we rename the filename by adding the timestamp
to the end of the filename. kludgy but works.
}
