% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_interval}
\alias{get_interval}
\title{Helper to regex the interval from the datapoints table}
\usage{
get_interval(string)
}
\arguments{
\item{string}{period string}
}
\value{
single character string
}
\description{
Helper to regex the interval from the datapoints table
}
