% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{get_legend_labels_from_df}
\alias{get_legend_labels_from_df}
\title{Get legend labels from input dataframe}
\usage{
get_legend_labels_from_df(df)
}
\arguments{
\item{df}{input dataframe with at least the following columns: `series_name`, `chart_no`}
}
\value{
character vector of length nrow(df) with legend labels
}
\description{
One of the columns is the `series_name` one, which is from the `series` table and contains
the dimension values separated by `--`. If these names have the same number of dimension
values and differ by just one of them, then that one is returned as the legend labels.
Warnings are issued for different numbers of dimensions or differences in more than
one dim, in which case NAs are returned.
If the labels have been entered manually, they are used instead - but must of course
be unique and not contain the `  -- ` character sequence.
At the moment no limits are placed on the length of the labels, that's coming next.
}
