% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{get_top_margin_and_title}
\alias{get_top_margin_and_title}
\title{Helper funciton to get the number of lines for the top margin and title position}
\usage{
get_top_margin_and_title(
  config,
  title_ps,
  legend_lead = 0.7,
  legend_offset = 0.25,
  title_gap = 0.3,
  title_lead = NULL
)
}
\arguments{
\item{config}{dictionary list from \link[UMARvisualisR]{prep_config}}

\item{title_ps}{title font size in points defaults 10}

\item{legend_lead}{controls spacing between legend entries}

\item{legend_offset}{controls spacing to plot area}
}
\value{
top margin in lines, vertical title position in lines and wrapped title
}
\description{
A bit of an eclectic function, but there's no way around it. This funciton
creates a blind plot to get the user dimensions
of the final chart, which lets it wrap the title, which lets it get the number
of lines for the title, which along with the number of lines in the legend
and a few graphical parameters lets us know what the top margin has to be.
Also returns the position of the title in lines and the wrapped title.
}
