% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub_ready_plots.R
\name{get_x_lims}
\alias{get_x_lims}
\title{Get x limits from the data}
\usage{
get_x_lims(datapoints)
}
\arguments{
\item{datapoints}{list of dataframes from \link[UMARvisualisR]{prep_data}}
}
\value{
date range
}
\description{
Used after the data has been cut to user xlims, this returns the
actual first and last date in the datapoints
}
