% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last_year_complete_series}
\alias{last_year_complete_series}
\title{Check if timeseries has a complete last year}
\usage{
last_year_complete_series(datapoints_df, max_year)
}
\arguments{
\item{datapoints_df}{dataframe with date column}

\item{max_year}{year to be checking if it is complete}
}
\value{
logical
}
\description{
But not if the max year is larger, then doesn't matter
}
