% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{left_axis_label_width}
\alias{left_axis_label_width}
\title{Prepare left axis labels and get width}
\usage{
left_axis_label_width(config, y_axis)
}
\arguments{
\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}

\item{y_axis}{output of \link[UMARvisualisR]{find_pretty_ylim}}
}
\value{
list of axis title, axis labels, and
number of lines
}
\description{
Changes euros to millions of euros if necessary and calculates how much
space the left margin needs to be. Returns axis title, axis labels, and
number of lines
}
