% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{left_axis_labels}
\alias{left_axis_labels}
\title{Draw left axis labels}
\usage{
left_axis_labels(unit, axis_positions, axis_labels, y_lab_lines)
}
\arguments{
\item{y_lab_lines}{number of margin lines from \link[UMARvisualisR]{left_axis_label_width}}

\item{config}{config dictionary list from \link[UMARvisualisR]{prep_config}}

\item{axes_labels}{labels from \link[UMARvisualisR]{left_axis_label_width}}
}
\value{
nothing, draws axis title and labels
}
\description{
Draw left axis labels
}
