% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check_n_prep.R
\name{multi_titles}
\alias{multi_titles}
\title{Check titles in input for multi-line chart}
\usage{
multi_titles(df)
}
\arguments{
\item{df}{input dataframe with at least the following columns: code, unit_name
interval_id, rolling_average_alignment, rolling_average_periods, year_on_year}
}
\value{
input df, possibly with updated main titles.
}
\description{
Check if the main_title is the same for all series, and if not issue a warning and
remove all the titles
}
