% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_special_barplot}
\alias{my_special_barplot}
\title{Barplot that can handle negative values}
\usage{
my_special_barplot(
  data_matrix,
  beside = FALSE,
  axes = TRUE,
  ylim = NULL,
  border = "white",
  col = NULL,
  panel.first = NULL,
  space = 0.2,
  xpd = TRUE,
  ...
)
}
\arguments{
\item{data_matrix}{vector or matrix of heights of the bars}

\item{...arguments}{to be passed to barplot}
}
\value{
nothing, just plots
}
\description{
A rework of the base barplot funciton that allows for negative
values to be properly plotted i.e. from zero downwards.
}
