% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{pdf_output}
\alias{pdf_output}
\title{Output to pdf}
\usage{
pdf_output(filename, size = "normal")
}
\arguments{
\item{filename}{name of file with path}

\item{size}{small, normal or large, defaults to normal}
}
\value{
nothing, produces pdf file
}
\description{
Opens device to output to pdf with filename and size. Afterwards you
need to call dev.off().
}
