% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{wrap_string}
\alias{wrap_string}
\title{Wrap a character string to a specified character width}
\usage{
wrap_string(string, width = 100, max_rows = 3)
}
\arguments{
\item{string}{a character string to be wrapped- vector of length one}

\item{width}{numeric - desired character width, defaults to 100}

\item{max_rows}{numeric maximum number of rows allowed in the output, defaults to 3}
}
\value{
a list with the first element a character string with new lines if required, and the
second one a numeric value for how many rows are in the returned string.
}
\description{
Uses strwrap to split a string into subsets that are shorter than the
specified width and inserts newlines in between them. Inspired by this
[SO answer](https://stackoverflow.com/questions/7367138/text-wrap-for-plot-titles/7367534#7367534).
Also returns the number of rows it has produced, and an optional max_row parameter
lets you set a threshold for the number of rows after which the text gets truncated.
}
